+++ src/main/java/org/sopt/user/dto/request/CreateUserRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.dto.request;
+package org.sopt.user.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
